% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcJaccardSimilarity.R
\name{calcJaccardSimilarity}
\alias{calcJaccardSimilarity}
\title{Calculate pairwise Jaccard similarity between all signatures}
\usage{
calcJaccardSimilarity(sets)
}
\arguments{
\item{sets}{a named list of signatures}
}
\value{
a Jaccard similarity matrix
}
\description{
Calculate pairwise Jaccard similarity between all signatures
}
\examples{
testlist <- list(a = 1:3, b = 3, c = 3:4)
jsim <- calcJaccardSimilarity(testlist)
}
