% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_contents.R
\name{getMostFrequentTaxa}
\alias{getMostFrequentTaxa}
\title{Get most frequent taxa}
\usage{
getMostFrequentTaxa(dat, n = 10, direction = c("BOTH", "UP", "DOWN"), ...)
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data}

\item{n}{integer. Number of taxa to show. Defaults to \code{10}
which will then show the 10 most frequent taxa.}

\item{direction}{character. Indicates direction of abundance change for signatures
to be included. Use \code{"UP"} to restrict
computation to signatures with increased abundance in the exposed group. Use 
\code{"DOWN"} to restrict to signatures with decreased abundance in the exposed
group. Defaults to \code{"BOTH"} which will not filter signatures by direction
of abundance change.}

\item{...}{Additional arguments passed on to \code{bugsigdbr::getSignatures}}
}
\value{
A sorted table listing absolute frequencies for the most frequent taxa
}
\description{
Get most frequent taxa
}
