% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSemanticSimilarity.R
\name{getNcbiTaxonomyObo}
\alias{getNcbiTaxonomyObo}
\title{Get NCBI Taxonomy in OBO format}
\usage{
getNcbiTaxonomyObo()
}
\value{
NCBI taxonomy in OBO format
}
\description{
Gets the NCBI taxonomy in OBO format.
}
