% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_contents.R
\name{getTaxLevelDistribution}
\alias{getTaxLevelDistribution}
\title{Get distribution of taxonomic levels}
\usage{
getTaxLevelDistribution(sigs)
}
\arguments{
\item{sigs}{a \code{list} of BugSigDB signatures in metaphlan format.
Typically obtained via \code{getSignatures}.}
}
\value{
A sorted table listing absolute frequencies of unique taxa for each
taxonomic level.
}
\description{
Get distribution of taxonomic levels
}
