% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_viz.R
\name{microbeHeatmap}
\alias{microbeHeatmap}
\title{Plot microbe co-occurrence in a heatmap}
\usage{
microbeHeatmap(
  dat,
  sig.type = c("both", "increased", "decreased"),
  tax.level = "mixed",
  exact.tax.level = FALSE,
  antagonistic = FALSE,
  anno = "phylum",
  top = 100,
  fontsize = 6,
  ...
)
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data.}

\item{sig.type}{character. Signature type. Use either \code{"increased"} or
\code{"decreased"} to subset to signatures with either increased or decreased abundance
in the exposed group, respectively. Default is \code{"both"} which will not 
subset by the direction of abundance change.}

\item{tax.level}{character. Either \code{"mixed"} or any subset of
\code{c("kingdom", "phylum", "class", "order", "family", "genus", "species",
"strain")}. This full vector is equivalent to \code{"mixed"}.}

\item{exact.tax.level}{logical. Should only the exact taxonomic level
specified by \code{tax.level} be returned? Defaults to \code{TRUE}.
If \code{FALSE}, a more general \code{tax.level} is extracted for
microbes given at a more specific taxonomic level.}

\item{antagonistic}{logical. Antagonistic co-occurrence, ie occurring together
but one microbe up the other one is down? Defaults to \code{FALSE}.}

\item{anno}{character. Taxonomic level that should be displayed as an annotation
bar. Defaults to \code{"phylum"}.}

\item{top}{integer. The number of microbes to display. Defaults to \code{100}.}

\item{fontsize}{integer. Fontsize for text. Defaults to \code{6}.}

\item{...}{additional arguments to \code{ComplexHeatmap::Heatmap}.}
}
\value{
Plots to a graphics device. Returns the co-oocurence matrix invisibly.
}
\description{
Plot microbe co-occurrence in a heatmap
}
