% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_viz.R
\name{plotComposition}
\alias{plotComposition}
\title{Plot composition of a table of a categorical variable}
\usage{
plotComposition(tab, column)
}
\arguments{
\item{tab}{a \code{table} with absolute frequencies of categories.}

\item{column}{character. The name of the variable for annotation of the plot.}
}
\value{
None. Plots to a graphics device.
}
\description{
Plot composition of a table of a categorical variable
}
