% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_viz.R
\name{plotProgressOverTime}
\alias{plotProgressOverTime}
\title{Plot curation output as a function of time}
\usage{
plotProgressOverTime(dat, col = "Curated date", diff = FALSE)
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data.}

\item{col}{character. A column of \code{dat} that contain the curation date
in dmy format.}

\item{diff}{logical. Display only the difference between months? Defaults to 
\code{FALSE} which will then display total cumulative numbers.}
}
\value{
None. Plots to a graphics device.
}
\description{
Plot curation output as a function of time
}
