% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_viz.R
\name{plotUniqueMicrobesOverTime}
\alias{plotUniqueMicrobesOverTime}
\title{Plot curation output for each curator}
\usage{
plotUniqueMicrobesOverTime(dat, date.col = "Curated date")
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data.}

\item{date.col}{character. A column of \code{dat} that contain the curation date
in dmy format.}
}
\value{
None. Plots to a graphics device.
}
\description{
Plot curation output for each curator
}
