% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_handling.R
\name{pmid2pubyear}
\alias{pmid2pubyear}
\title{Get the publication year for a PMID}
\usage{
pmid2pubyear(pmid)
}
\arguments{
\item{pmid}{character. A PubMed ID.}
}
\value{
A character containg the publication year for the given \code{pmid}.
}
\description{
Get the publication year for a PMID
}
\examples{
 pmid2pubyear("32026945")
}
