% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_contents.R
\name{sigStat}
\alias{sigStat}
\title{Signature stats}
\usage{
sigStat(dat, col)
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data}

\item{col}{character. A column of \code{dat} that should be tabled.}
}
\value{
A sorted table with signature stats for the chosen metadata column
}
\description{
Signature stats
}
