% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_handling.R
\name{stratifyByCurator}
\alias{stratifyByCurator}
\title{Stratify curation output by curator}
\usage{
stratifyByCurator(dat)
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data}
}
\value{
a numeric \code{matrix} storing number of papers and number of 
signatures curated for each curator
}
\description{
Stratify curation output by curator
}
