% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_handling.R
\name{stripEmptySignatures}
\alias{stripEmptySignatures}
\title{Remove entries that do not contain signatures}
\usage{
stripEmptySignatures(dat, col = "NCBI Taxonomy IDs")
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data}

\item{col}{character. A column of \code{dat} that contain signatures
as a comma separated list}
}
\value{
A \code{data.frame} with empty signature entries removed.
}
\description{
Remove entries that do not contain signatures
}
