% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSemanticSimilarity.R
\name{syncWithNCBI}
\alias{syncWithNCBI}
\title{Synchronize signatures with NCBI Taxonomy}
\usage{
syncWithNCBI(sigs, onto)
}
\arguments{
\item{sigs}{signatures. A list of character vectors. Typically obtained 
via \code{\link{getSignatures}}.}

\item{onto}{ontology. An object of class \code{ontology_index} storing the 
NCBI Taxonomy. Typically obtained via \code{\link{getNcbiTaxonomyObo}}.}
}
\value{
The input signatures synchronized with the NCBI Taxonomy
}
\description{
Helper function for synchronizing signatures with NCBI Taxonomy
}
\examples{
 library(bugsigdbr)
 df <- importBugSigDB()
 sigs <- getSignatures(df)
 onto <- getNcbiTaxonomyObo()
 sigs <- syncWithNCBI(sigs, onto)
 
}
