% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_contents.R
\name{tabCol}
\alias{tabCol}
\title{Table column}
\usage{
tabCol(df, col, n = Inf, perc = FALSE)
}
\arguments{
\item{df}{a \code{data.frame} storing BugSigDB data}

\item{col}{character. A column of \code{dat} that should be tabled.}

\item{n}{integer. Number of categories to show. Defaults to \code{Inf}
which will then show all categories in the chosen column.}

\item{perc}{logical. Return absolute frequencies or relative frequencies?
Defaults to \code{FALSE} which will return absolute frequencies.}
}
\value{
A sorted table with table stats for the chosen metadata column
}
\description{
Table column
}
