% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_contents.R
\name{tabDiv}
\alias{tabDiv}
\title{Table diversity column}
\usage{
tabDiv(exps, div.col, spl.col, min.exps = 5, perc = FALSE)
}
\arguments{
\item{exps}{a \code{data.frame} storing BugSigDB experiment data}

\item{div.col}{character. A column of \code{exps} that should be tabled.}

\item{spl.col}{character. A column of \code{exps} that should be used to
split \code{div.col} into groups.}

\item{min.exps}{integer. Minimum number of experiments for a category in
\code{spl.col} to be included in the result. Defaults to \code{5}.}

\item{perc}{logical. Return absolute frequencies or relative frequencies?
Defaults to \code{FALSE} which will return absolute frequencies.}
}
\value{
A sorted table with table stats for the chosen diversity column
}
\description{
Table diversity column
}
