% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_contents.R
\name{testAssociation}
\alias{testAssociation}
\title{Test association of a taxon with a certain category of a column}
\usage{
testAssociation(
  dat,
  taxon,
  column = "Condition",
  category = "",
  direction = c("BOTH", "UP", "DOWN")
)
}
\arguments{
\item{dat}{a \code{data.frame} storing BugSigDB data}

\item{taxon}{character. A taxonomic name.}

\item{column}{character. A column of \code{dat} that contains the category 
thats hould be tested for association.}

\item{category}{character. One or more categories that should be tested for
association.}

\item{direction}{character. Indicates direction of abundance change for signatures
to be included. Use \code{"UP"} to restrict
computation to signatures with increased abundance in the exposed group. Use 
\code{"DOWN"} to restrict to signatures with decreased abundance in the exposed
group. Defaults to \code{"BOTH"} which will not filter signatures by direction
of abundance change.}
}
\value{
A sorted table with taxon stats for the chosen metadata column
}
\description{
Test association of a taxon with a certain category of a column
}
