% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSemanticSimilarity.R
\name{weightedBMA}
\alias{weightedBMA}
\title{Weighted semantic similarity}
\usage{
weightedBMA(
  o,
  ic,
  set1,
  set2,
  weights1 = rep(1, length(set1)),
  weights2 = rep(1, length(set2))
)
}
\arguments{
\item{o}{ontology. An object of class \code{ontology_index}.}

\item{ic}{information content. Typically obtained via 
\code{ontologySimilarity::descendants_IC(o)}.}

\item{set1}{character. First term set.}

\item{set2}{character. Secound term set.}

\item{weights1}{numeric. Weights for each term of the first term set, ie. must
be parallel to \code{set1}.}

\item{weights2}{numeric. Weights for each term of the second term set, ie. must
be parallel to \code{set2}.}
}
\value{
a numeric value expressing semantic similarity between the two
input term sets, weighted by the given individual term weights.
}
\description{
Incorporation of weights into the computation of semantic similarity
}
