% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSemanticSimilarity.R
\name{calcSemanticSimilarity}
\alias{calcSemanticSimilarity}
\title{Calculate semantic similarity}
\usage{
calcSemanticSimilarity(file1, file2)
}
\arguments{
\item{file1}{Signature exported from BugSigDB.org.}

\item{file2}{Signature exported from BugSigDB.org.}
}
\description{
\code{calcSemanticSimilarity} calculates the pairwise semantic similarity
between two sets of signature files exported from BugSigDB.org.
}
\examples{

\dontrun{
file1 <- "path/to/exported/file1.csv"
file2 <- "path/to/exported/file2.csv"
output <- calcSemanticSimilarity(file1, file2)
head(output)
}

}
